Uncomment TrustHosts middleware to enable it by default
We recently learned of a vulnerability in our app through our Bug Bounty program which allowed an attacker to change the host of a signed URL using the X-Forwarded-Host header. This allowed an attacker to generate a malicious password-reset email which, if the link was clicked, would send the user's password reset token to a website controlled by the hacker.

Enabling this middleware by default could close this vulnerability for many websites who may currently be exposed to this same vulnerability.

Signed-off-by: Salim Djerbouh <caddydz4@gmail.com>