Add date facade alias
All other facades have an alias (except for ParallelTesting - which I don't think needs one).

I think having an alias for this makes sense as it can by handy in views to be able to work with dates that aren't always coming from a model where a cast is usable and an alias makes this nicer as we don't need to import the namespace. It is also handy in tinker to just call `Date::parse()` to test some stuff on the fly.

Signed-off-by: Salim Djerbouh <caddydz4@gmail.com>